Add statz endpoint.
