+++ packages/zero-cache/package.json
@@ -37,6 +37,8 @@
     "@rocicorp/logger": "^5.4.0",
     "@rocicorp/resolver": "^1.0.2",
     "@rocicorp/zero-sqlite3": "^1.0.4",
+    "@types/basic-auth": "^1.1.8",
+    "basic-auth": "^2.0.1",
     "chalk": "^5.3.0",
     "compare-utf8": "^0.1.1",
     "defu": "^6.1.4",
