+++ packages/zero-cache/src/config/zero-config.ts
@@ -341,6 +341,14 @@ export const zeroOptions = {
     ],
   },
 
+  adminPassword: {
+    type: v.string().optional(),
+    desc: [
+      `A password used to administer zero-cache server, for example to access the`,
+      `/statz endpoint.`,
+    ],
+  },
+
   litestream: {
     executable: {
       type: v.string().optional(),
@@ -532,16 +540,16 @@ export const zeroOptions = {
   },
 };
 
-export type ZeroConfig = Config<typeof zeroOptions>;
+export type ZeroConfig = ReturnType<typeof getZeroConfig>;
 
 export const ZERO_ENV_VAR_PREFIX = 'ZERO_';
 
-let loadedConfig: ZeroConfig | undefined;
+let loadedConfig: Config<typeof zeroOptions> | undefined;
 
 export function getZeroConfig(
   env: NodeJS.ProcessEnv = process.env,
   argv = process.argv.slice(2),
-): ZeroConfig {
+) {
   if (!loadedConfig || singleProcessMode()) {
     loadedConfig = parseOptions(zeroOptions, argv, ZERO_ENV_VAR_PREFIX, env);
 
@@ -550,5 +558,15 @@ export function getZeroConfig(
     }
   }
 
-  return loadedConfig;
+  return {
+    ...loadedConfig,
+    cvr: {
+      ...loadedConfig.cvr,
+      db: loadedConfig.cvr.db ?? loadedConfig.upstream.db,
+    },
+    change: {
+      ...loadedConfig.change,
+      db: loadedConfig.change.db ?? loadedConfig.upstream.db,
+    },
+  };
 }
