+++ packages/zero-cache/src/server/change-streamer.ts
@@ -103,6 +103,7 @@ export default async function runWorker(
   assert(changeStreamer, `resetting replica did not advance replicaVersion`);
 
   const changeStreamerWebServer = new ChangeStreamerHttpServer(
+    config,
     lc,
     changeStreamer,
     {port},
