+++ packages/zero-cache/src/services/change-streamer/change-streamer-http.test.ts
@@ -24,6 +24,7 @@ import {
 } from './change-streamer-http.ts';
 import type {Downstream, SubscriberContext} from './change-streamer.ts';
 import {PROTOCOL_VERSION} from './change-streamer.ts';
+import type {ZeroConfig} from '../../config/zero-config.ts';
 
 describe('change-streamer/http', () => {
   let lc: LogContext;
@@ -47,17 +48,26 @@ describe('change-streamer/http', () => {
 
     const [parent, receiver] = inProcChannel();
 
-    dispatcher = new HttpService('dispatcher', lc, {port: 0}, fastify => {
-      installWebSocketHandoff(
-        lc,
-        req => ({payload: getSubscriberContext(req), receiver}),
-        fastify.server,
-      );
-    });
+    const config = {} as unknown as ZeroConfig;
+
+    dispatcher = new HttpService(
+      'dispatcher',
+      config,
+      lc,
+      {port: 0},
+      fastify => {
+        installWebSocketHandoff(
+          lc,
+          req => ({payload: getSubscriberContext(req), receiver}),
+          fastify.server,
+        );
+      },
+    );
 
     // Run the server for real instead of using `injectWS()`, as that has a
     // different behavior for ws.close().
     server = new ChangeStreamerHttpServer(
+      config,
       lc,
       {subscribe: subscribeFn.mockResolvedValue(downstream)},
       {port: 0},
