+++ packages/zero-cache/src/services/change-streamer/change-streamer-http.ts
@@ -16,6 +16,7 @@ import {
   type Downstream,
   type SubscriberContext,
 } from './change-streamer.ts';
+import type {ZeroConfig} from '../../config/zero-config.ts';
 
 const MIN_SUPPORTED_PROTOCOL_VERSION = 1;
 
@@ -31,12 +32,13 @@ export class ChangeStreamerHttpServer extends HttpService {
   readonly #delegate: ChangeStreamer;
 
   constructor(
+    config: ZeroConfig,
     lc: LogContext,
     delegate: ChangeStreamer,
     opts: Options,
     parent: Worker,
   ) {
-    super('change-streamer-http-server', lc, opts, async fastify => {
+    super('change-streamer-http-server', config, lc, opts, async fastify => {
       await fastify.register(websocket);
 
       // fastify does not support optional path components, so we just
