--- packages/zero/src/cache-stats.ts
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-import '../../analyze-query/src/bin-cache-stats.ts';
